feat(cross-check): add empty page state
